requirements.txt
flask>=3.0.0
werkzeug>=3.0.0
python-dotenv>=1.0.0
anthropic>=0.40.0
flask-limiter>=3.5.0
Pillow>=10.0.0
flask-session>=0.8.0
flask-wtf>=1.2.0
